Trait: CheckHelper
