+++ src/DeploymentDetails.php
@@ -4,6 +4,8 @@ namespace Marcz\Phar\NautPie;
 
 class DeploymentDetails
 {
+    use CheckHelper;
+
     protected $details = [
         'ref' => '',
         'ref_type' => 'sha',
@@ -86,34 +88,5 @@ class DeploymentDetails
         $this->details['ref_type'] = $type;
 
         return $this;
-    }
-
-    /**
-     * Work-around to use isNotNull (aka isset)  as a callback function
-     * @param  mixed $value
-     * @return boolean
-     */
-    public function isNotNull($value)
-    {
-        // Note: Because this is a language construct and not a function,
-        // it cannot be called using variable functions.
-        return isset($value);
-    }
-
-    public function checkBoolean($value)
-    {
-        if (in_array(strtolower($value), ['false', 'no', '0'], 1)) {
-            $value = false;
-        }
-
-        if (in_array(strtolower($value), ['true', 'yes', '1'], 1)) {
-            $value = true;
-        }
-
-        if (!is_bool($value)) {
-            throw new \Exception('[Boolean] Value not accepted as boolean', 1);
-        }
-
-        return (bool) $value;
     }
 }
